Refactor EH filter structuring
